language france and geolocalisation enabled (still in debug)

+++ app/src/main/java/thermo/aziaka/donavan/com/thermo/POJO/Main.java
@@ -6,7 +6,7 @@ package thermo.aziaka.donavan.com.thermo.POJO;
 
 public class Main {
     private float temp;
-    private int pressure;
+    private float pressure;
     private int humidity;
     private float temp_min;
     private float temp_max;
@@ -19,7 +19,7 @@ public class Main {
         this.temp = temp;
     }
 
-    public int getPressure() {
+    public float getPressure() {
         return pressure;
     }
