+++ app/src/main/java/thermo/aziaka/donavan/com/thermo/POJO/Wind.java
@@ -6,7 +6,7 @@ package thermo.aziaka.donavan.com.thermo.POJO;
 
 public class Wind {
     private float speed;
-    private int deg;
+    private float deg;
 
     public float getSpeed() {
         return speed;
@@ -16,7 +16,7 @@ public class Wind {
         this.speed = speed;
     }
 
-    public int getDeg() {
+    public float getDeg() {
         return deg;
     }
